requirements.txt
numpy
pandas
keras==2.12.0
tensorflow==2.12.0
opencv-python
typing_extensions
Pillow
matplotlib
scikit-learn
imutils
streamlit